Iter23 (#24)
Iter23

+++ internal/sync/semaphore.go
@@ -6,18 +6,18 @@ type Semaphore struct {
 }
 
 // NewSemaphore создает семафор с буферизованным каналом емкостью maxReq
-func NewSemaphore(maxReq uint) *Semaphore {
+func NewSemaphore(maxReq uint64) *Semaphore {
 	return &Semaphore{
 		semaCh: make(chan struct{}, maxReq),
 	}
 }
 
-// когда горутина запускается, отправляем пустую структуру в канал semaCh
+// Acquire когда горутина запускается, отправляем пустую структуру в канал semaCh.
 func (s *Semaphore) Acquire() {
 	s.semaCh <- struct{}{}
 }
 
-// когда горутина завершается, из канала semaCh убирается пустая структура
+// Release когда горутина завершается, из канала semaCh убирается пустая структура
 func (s *Semaphore) Release() {
 	<-s.semaCh
 }
